Fixed compilation errors

+++ src/BlockEntities/BrewingstandEntity.cpp
@@ -61,7 +61,7 @@ void cBrewingstandEntity::CopyFrom(const cBlockEntity & a_Src)
 	Super::CopyFrom(a_Src);
 	auto & src = reinterpret_cast<const cBrewingstandEntity &>(a_Src);
 	m_IsBrewing = src.m_IsBrewing;
-	m_CurrentBrewingRecipes = src.m_CurrentBrewingRecipes;
+    m_CurrentBrewingRecipes = src.m_CurrentBrewingRecipes;
 	m_Results = src.m_Results;
 	m_TimeBrewed = src.m_TimeBrewed;
 	m_RemainingFuel = src.m_RemainingFuel;
@@ -131,12 +131,12 @@ bool cBrewingstandEntity::Tick(std::chrono::milliseconds a_Dt, cChunk & a_Chunk)
 		const cBrewingRecipes::cRecipe * Recipe = nullptr;
 		for (int i = 0; i < 3; i++)
 		{
-			if (m_Contents.GetSlot(i).IsEmpty() || (m_CurrentBrewingRecipes[i] == nullptr))
+            if (m_Contents.GetSlot(i).IsEmpty() || (m_CurrentBrewingRecipes[static_cast<size_t>(i)] == nullptr))
 			{
 				continue;
 			}
 
-			Recipe = m_CurrentBrewingRecipes[i];
+            Recipe = m_CurrentBrewingRecipes[static_cast<size_t>(i)];
 			m_Contents.SetSlot(i, Recipe->Output.CopyOne());
 		}
 
@@ -254,14 +254,14 @@ void cBrewingstandEntity::OnSlotChanged(cItemGrid * a_ItemGrid, int a_SlotNum)
 	{
 		if (GetSlot(i).IsEmpty())
 		{
-			m_CurrentBrewingRecipes[i] = nullptr;
-			m_Results[i].Clear();
+            m_CurrentBrewingRecipes[static_cast<size_t>(i)] = nullptr;
+            m_Results[static_cast<size_t>(i)].Clear();
 			continue;
 		}
 
-		if (m_CurrentBrewingRecipes[i] != nullptr)
+        if (m_CurrentBrewingRecipes[static_cast<size_t>(i)] != nullptr)
 		{
-			Recipe = m_CurrentBrewingRecipes[i];
+            Recipe = m_CurrentBrewingRecipes[static_cast<size_t>(i)];
 			if (Recipe->Ingredient.IsEqual(GetSlot(bsIngredient)) && Recipe->Input.IsEqual(GetSlot(i)))
 			{
 				Stop = false;
@@ -273,8 +273,8 @@ void cBrewingstandEntity::OnSlotChanged(cItemGrid * a_ItemGrid, int a_SlotNum)
 		if (Recipe != nullptr)
 		{
 			// Found a brewing recipe for the items
-			m_CurrentBrewingRecipes[i] = Recipe;
-			m_Results[i] = Recipe->Output.CopyOne();
+            m_CurrentBrewingRecipes[static_cast<size_t>(i)] = Recipe;
+            m_Results[static_cast<size_t>(i)] = Recipe->Output.CopyOne();
 			Stop = false;
 		}
 	}
@@ -348,8 +348,8 @@ void cBrewingstandEntity::LoadRecipes(void)
 		Recipe = BR->GetRecipeFrom(GetSlot(i), GetSlot(bsIngredient));
 		if (Recipe != nullptr)
 		{
-			m_CurrentBrewingRecipes[i] = Recipe;
-			m_Results[i] = Recipe->Output.CopyOne();
+            m_CurrentBrewingRecipes[static_cast<size_t>(i)] = Recipe;
+            m_Results[static_cast<size_t>(i)] = Recipe->Output.CopyOne();
 		}
 	}
 }
